Update main application
The application now has support for command line arguments. In fact, it
is required. Users have three options:
    Display a help message with --help or -h
    Specify a filepath with --path or -p. This is a required argument.
    Specify a size for the histogram with --size or -s.
	This is a required argument.

If the filepath or size argument are not specified, the program will
throw an exception. An exception is thrown if the path provided does not
exist. A helper print statement is provided.

The regex statement has been changed to match the pattern of a number
rather than a specific range of numbers.